requirements.txt
flask
torch
torchvision
segmentation-models-pytorch
albumentations
opencv-python
pillow
numpy
